chore(fileio): drop redundant i64→i64 casts in cost_spec
Post-A8-M-1 the FS_ENTRIES_SETUP / FS_ENTRIES_PER_ENTRY constants
in `costs.rs` were retyped from `u64` to `i64` (matching the return
type of `.saturating_add` on `Cost::value`).  The `as i64` casts at
the two call sites in fileio_cost_spec.rs became identity casts, and
clippy (correctly) flags them.

No behavior change; unblocks the pre-commit `-D warnings` gate on
this branch.

+++ rholang/tests/fileio_cost_spec.rs
@@ -474,9 +474,8 @@ fn entries_family_supplements_match_combined_costs() {
         // combined weight must equal `FS_ENTRIES_SETUP + n *
         // FS_ENTRIES_PER_ENTRY` under saturation (matches the retired
         // bulk helper's shape).
-        let expected_next = (FS_ENTRIES_SETUP as i64).saturating_add(
-            (FS_ENTRIES_PER_ENTRY as i64).saturating_mul(n.min(i64::MAX as u64) as i64),
-        );
+        let expected_next = FS_ENTRIES_SETUP
+            .saturating_add(FS_ENTRIES_PER_ENTRY.saturating_mul(n.min(i64::MAX as u64) as i64));
         assert_eq!(
             fs_entries_stream_next_cost()
                 .value
